+++ internal/provider/lxc/statics.go
@@ -1,0 +1,8 @@
+package lxc
+
+const DESC_RSRC_LXC_EXEC = `lxc_exec execute commands inside an lxc.
+
+- If a command fails, the execution will stop and the resource will not be created.`
+const MD_RSRC_LXC_EXEC = `**lxc_exec** execute commands inside an lxc.
+
+- If a command fails, the execution will stop and the resource will not be created.`